+++ settings.php
@@ -24,17 +24,36 @@ defined('MOODLE_INTERNAL') || die;
 
 if ($ADMIN->fulltree) {
 
-    // Toggle all on/off.
-    // 1 => No.
-    // 2 => Yes.
-
-    $name = 'block_course_modulenavigation/toggleclickonarrow';
-    $title = get_string('toggleclickonarrow', 'block_course_modulenavigation');
-    $description = get_string('toggleclickonarrow_desc', 'block_course_modulenavigation');
-    $default = 2;
+    // Option: clicking on the downwards arrow 1) displays the menu or 2)goes to that page
+    $name = 'block_course_modulenavigation/toggleclickontitle';
+    $title = get_string('toggleclickontitle', 'block_course_modulenavigation');
+    $description = get_string('toggleclickontitle_desc', 'block_course_modulenavigation');
+    $default = 1;
     $choices = array(
-        1 => get_string('toggleclickonarrow_menu', 'block_course_modulenavigation'),
-        2 => get_string('toggleclickonarrow_page', 'block_course_modulenavigation'),
+        1 => get_string('toggleclickontitle_menu', 'block_course_modulenavigation'),
+        2 => get_string('toggleclickontitle_page', 'block_course_modulenavigation'),
+    );
+    $settings->add(new admin_setting_configselect($name, $title, $description, $default, $choices));
+
+    // Option: show labels
+    $name = 'block_course_modulenavigation/toggleshowlabels';
+    $title = get_string('toggleshowlabels', 'block_course_modulenavigation');
+    $description = get_string('toggleshowlabels_desc', 'block_course_modulenavigation');
+    $default = 1;
+    $choices = array(
+        1 => new lang_string('no'), // No.
+        2 => new lang_string('yes')   // Yes.
+    );
+    $settings->add(new admin_setting_configselect($name, $title, $description, $default, $choices));
+
+    // Option: Show all tabs open
+    $name = 'block_course_modulenavigation/togglecollapse';
+    $title = get_string('togglecollapse', 'block_course_modulenavigation');
+    $description = get_string('togglecollapse_desc', 'block_course_modulenavigation');
+    $default = 1;
+    $choices = array(
+        1 => new lang_string('no'), // No.
+        2 => new lang_string('yes')   // Yes.
     );
     $settings->add(new admin_setting_configselect($name, $title, $description, $default, $choices));
 }