+++ changes.md
@@ -17,6 +17,11 @@ Version 3.2 (2016.06.23)
   3. Fix a bug with the option "show the current section only" where link of title section doesn't work
 
 
-Version 3.2 (2016.08.01)
+Version 3.3 (2016.08.01)
   1. Add links on the sections titles. Now each section title point to section page
-  2. Add an arrow after sections titles to open menu content. 
+  2. Add an arrow after sections titles to open menu content.
+
+Version 3.4 (2016.10.07)
+  1. Add option between downwards arrow and click on title
+  2. Show labels in the menu
+  3. Add option to show all tabs open